Spliting main.js to modules
