+++ .gitignore
@@ -13,3 +13,7 @@ node_modules/
 
 # Папка с собранными файлами проекта
 build/
+
+# Собранные JS-файлы (бандл)
+public/bundle.js
+public/bundle.js.map
