+++ public/index.html
@@ -28,6 +28,6 @@
   </section>
 </footer>
 
-<script src="../src/main.js"></script>
+<script src="./bundle.js"></script>
 </body>
 </html>
