+++ src/main.js
@@ -1,113 +1,12 @@
-"use strict";
-
-const FILM_CARDS_AMOUNT = 5;
-const EXTRA_FILM_CARDS_AMOUNT = 2;
+import createUserProfileTemplate from "./view/profile.js";
+import createNavigationTemplate from "./view/navigation.js";
+import createFiltersTemplate from "./view/filters.js";
+import createFilmSectionTemplate from "./view/film-board.js";
+import createStatsTemplate from "./view/stats.js";
 
 const mainHeaderElement = document.querySelector(`.header`);
 const mainElement = document.querySelector(`.main`);
 const mainFooterElement = document.querySelector(`.footer`);
-
-const createUserProfileTemplate = () => {
-  return (
-    `<section class="header__profile profile">
-      <p class="profile__rating">Movie Buff</p>
-      <img
-        class="profile__avatar"
-        src="images/bitmap@2x.png"
-        alt="Avatar"
-        width="35"
-        height="35"
-      >
-    </section>`
-  );
-};
-
-const createNavigationTemplate = () => {
-  return (
-    `<nav class="main-navigation">
-      <div class="main-navigation__items">
-        <a href="#all" class="main-navigation__item main-navigation__item--active">All movies</a>
-        <a href="#watchlist" class="main-navigation__item">Watchlist <span class="main-navigation__item-count">13</span></a>
-        <a href="#history" class="main-navigation__item">History <span class="main-navigation__item-count">4</span></a>
-        <a href="#favorites" class="main-navigation__item">Favorites <span class="main-navigation__item-count">8</span></a>
-      </div>
-      <a href="#stats" class="main-navigation__additional">Stats</a>
-    </nav>`
-  );
-};
-
-const createFiltersTemplate = () => {
-  return (
-    `<ul class="sort">
-      <li><a href="#" class="sort__button sort__button--active">Sort by default</a></li>
-      <li><a href="#" class="sort__button">Sort by date</a></li>
-      <li><a href="#" class="sort__button">Sort by rating</a></li>
-    </ul>`
-  );
-};
-
-const createFilmCardTemplate = () => {
-  return (
-    `<article class="film-card">
-      <h3 class="film-card__title">The Dance of Life</h3>
-      <p class="film-card__rating">8.3</p>
-      <p class="film-card__info">
-        <span class="film-card__year">1929</span>
-        <span class="film-card__duration">1h 55m</span>
-        <span class="film-card__genre">Musical</span>
-      </p>
-      <img src="./images/posters/the-dance-of-life.jpg" alt="" class="film-card__poster">
-      <p class="film-card__description">Burlesque comic Ralph "Skid" Johnson (Skelly), and specialty dancer Bonny Lee King (Carroll), end up together on a cold, rainy night at a tr…</p>
-      <a class="film-card__comments">5 comments</a>
-      <form class="film-card__controls">
-        <button class="film-card__controls-item button film-card__controls-item--add-to-watchlist">Add to watchlist</button>
-        <button class="film-card__controls-item button film-card__controls-item--mark-as-watched">Mark as watched</button>
-        <button class="film-card__controls-item button film-card__controls-item--favorite">Mark as favorite</button>
-      </form>
-    </article>`
-  );
-};
-
-const createExtraFilmListTemplate = (heading, cardAamount) => {
-  return (
-    `<section class="films-list--extra">
-      <h2 class="films-list__title">${heading}</h2>
-
-      <div class="films-list__container">
-        ${[...new Array(cardAamount)]
-          .map(() => createFilmCardTemplate())
-          .join(``)}
-      <div>
-    </section>`
-  );
-};
-
-const createFilmSectionTemplate = () => {
-  return (
-    `<section class="films">
-      <section class="films-list">
-        <h2 class="films-list__title visually-hidden">All movies. Upcoming</h2>
-
-        <div class="films-list__container">
-          ${[...new Array(FILM_CARDS_AMOUNT)]
-            .map(() => createFilmCardTemplate())
-            .join(``)}
-        </div>
-        <button class="films-list__show-more">Show more</button>
-      </section>
-      ${createExtraFilmListTemplate(`Top rated`, EXTRA_FILM_CARDS_AMOUNT)}
-      ${createExtraFilmListTemplate(`Most commented`, EXTRA_FILM_CARDS_AMOUNT)}
-    `
-  );
-};
-
-const createStatsTemplate = () => {
-  return (
-    `<section class="footer__statistics">
-      <p>130 291 movies inside</p>
-    </section>`
-  );
-};
 
 const renderTemplate = (container, template) => {
   container.insertAdjacentHTML(`beforeend`, template);
