+++ src/view/stats.js
@@ -1,0 +1,9 @@
+const createStatsTemplate = () => {
+  return (
+    `<section class="footer__statistics">
+      <p>130 291 movies inside</p>
+    </section>`
+  );
+};
+
+export default createStatsTemplate;
